universities-service/requirements.txt
fastapi
uvicorn[standard]
requests
pymongo
python-dotenv
pydantic
pydantic[email]
pytz
httpx
